expand to cover a wider variety of clients
Deal with clients:
that have been previously registered to a different server
suffering from bugzilla 1422518
RHEL as well as previously CentOS
where standard EPEL installation by package fails

Keep a record of channel list and orphaned packages both locally and on
the server the first time this is run.
Push keys over and add them with copy and rpm_key.